backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
pydantic==2.10.3
PyMuPDF==1.25.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
pytesseract==0.3.13
pillow==11.1.0
easyocr==1.7.2
sentence-transformers==3.4.0
faiss-cpu==1.13.2
openai==1.57.0
python-dotenv==1.0.1
groq
pydantic-settings>=2.0.0
alembic>=1.13.0
